feat(payment): 接入api-用户管理

+++ system/service/src/main/java/com/xh/business/domain/resp/user/UserVO.java
@@ -35,10 +35,11 @@ public class UserVO implements Serializable {
      * 账号状态（0- 正常 1- 封号）
      */
     private Integer status;
+
     /**
      * 钱包余额（分）
      */
-    private Integer balance;
+    private Long balance;
 
     /**
      * 账号
